+++ packages/react/src/checkbox/checkbox-control.tsx
@@ -6,10 +6,15 @@ import { useCheckboxContext } from './checkbox-context'
 export type CheckboxControlProps = ComponentPropsWithoutRef<typeof ark.div>
 
 export const CheckboxControl = forwardRef<HTMLDivElement, CheckboxControlProps>((props, ref) => {
-  const { controlProps } = useCheckboxContext()
+  const { controlProps, inputProps } = useCheckboxContext()
   const mergedProps = mergeProps(controlProps, props)
 
-  return <ark.div {...mergedProps} ref={ref} />
+  return (
+    <>
+      <ark.div {...mergedProps} ref={ref} />
+      <input {...inputProps} />
+    </>
+  )
 })
 
 CheckboxControl.displayName = 'CheckboxControl'