+++ packages/vue/src/checkbox/stories/basic.stories.vue
@@ -1,15 +1,11 @@
 <script setup lang="ts">
-import { ref } from 'vue'
-import { Checkbox, CheckboxControl, CheckboxInput, CheckboxLabel } from '..'
+import { Checkbox, CheckboxControl, CheckboxLabel } from '..'
 import '../checkbox.css'
-
-const checked = ref(true)
-const props = defineProps(['modelValue'])
 </script>
+
 <template>
   <Checkbox>
     <CheckboxLabel>Checkbox</CheckboxLabel>
-    <CheckboxInput />
     <CheckboxControl />
   </Checkbox>
 </template>